+++ Domain/EntitiesForManagement/Building.cs
@@ -18,8 +18,8 @@ public class Building
     public string Description { get; set; }
     public int TotalRooms { get; set; }
     public string ImageUrl { get; set; }
-    public double CoordinateX { get; set; }
-    public double CoordinateY { get; set; }
+    public decimal CoordinateX { get; set; }
+    public decimal CoordinateY { get; set; }
 
     public bool Status { get; set; }
