+++ Domain/EntityRequest/Building/BuildingUpdateRequest.cs
@@ -4,9 +4,9 @@ public class BuildingUpdateRequest
 {
     public string BuildingName { get; set; } = null!;
     public string Description { get; set; } = null!;
-    public double? CoordinateX { get; set; }
+    public decimal? CoordinateX { get; set; }
     public string? ImageUrl { get; set; }
-    public double? CoordinateY { get; set; }
+    public decimal? CoordinateY { get; set; }
     public bool Status { get; set; }
     public int AreaId { get; set; }
 }
