+++ Domain/EntityRequest/Request/TicketCreateRequest.cs
@@ -5,7 +5,7 @@ public class TicketCreateRequest
     public string RequestName { get; set; } = null!;
     public string Description { get; set; } = null!;
     public DateTime? SolveDate { get; set; }
-    public double? Amount { get; set; }
+    public decimal? Amount { get; set; }
     public string Status { get; set; } = null!;
     public int TicketTypeId { get; set; }
     public int RenterId { get; set; }
